COC-130-refactor(#20): IconButton 컴포넌트 리펙토링
