+++ src/components/_common/atoms/Tag/index.tsx
@@ -3,6 +3,6 @@ import S, { TagStyleProps } from './style';
 
 export type TagProps = PropsWithChildren<React.ComponentProps<'div'> & TagStyleProps>;
 
-export default function Tag({ children, color }: TagProps) {
+export default function Tag({ children, color = 'gray' }: TagProps) {
   return <S.Tag color={color}>{children}</S.Tag>;
 }
